Widget spacing ui fix
